Extract msi tests in a new file
